change to i32

+++ src/main.rs
@@ -1,4 +1,4 @@
-fn unique(a: Vec<f32>) -> Vec<f32> {
+fn unique(a: Vec<i32>) -> Vec<i32> {
     todo!();
 }
 
@@ -18,7 +18,7 @@ fn unique(a: Vec<f32>) -> Vec<f32> {
 // }
 
 fn main() {
-    let input = vec![2.0, 1.0, 1.0];
+    let input = vec![2, 1, 1];
     let answer = unique(input);
     println!("unique items -> {:?}", answer)
 }
@@ -34,16 +34,16 @@ fn empty_list() {
 
 #[test]
 fn sorted_list() {
-    let input = vec![1.0, 4.0, 5.0];
-    let expected_output = vec![1.0, 4.0, 5.0];
+    let input = vec![1, 4, 5];
+    let expected_output = vec![1, 4, 5];
     let actual_output = unique(input);
     assert_eq!(actual_output, expected_output);
 }
 
 #[test]
 fn unsorted_list() {
-    let input = vec![1.0, 5.0, 2.0];
-    let expected_output = vec![1.0, 5.0, 2.0];
+    let input = vec![1, 5, 2];
+    let expected_output = vec![1, 5, 2];
     let actual_output = unique(input);
     assert_eq!(actual_output, expected_output);
 }
@@ -51,17 +51,17 @@ fn unsorted_list() {
 
 #[test]
 fn unsorted_list_with_duplicates() {
-    let input = vec![1.0, 5.0, 2.0, 2.0, 1.0];
-    let expected_output = vec![1.0, 5.0, 2.0];
+    let input = vec![1, 5, 2, 2, 1];
+    let expected_output = vec![1, 5, 2];
     let actual_output = unique(input);
     assert_eq!(actual_output, expected_output);
 }
 
 #[test]
 fn sorted_list_with_duplicates() {
-    let mut input = vec![1.0, 5.0, 2.0, 2.0, 1.0];
+    let mut input = vec![1, 5, 2, 2, 1];
     input.sort_by(|x,y| x.partial_cmp(y).unwrap());
-    let expected_output = vec![1.0, 2.0, 5.0];
+    let expected_output = vec![1, 2, 5];
     let actual_output = unique(input);
     assert_eq!(actual_output, expected_output);
 }
